refactor based on google code style
