+++ app/api/doc/doc_urls.py
@@ -3,16 +3,15 @@
 from django.urls import path
 from rest_framework.schemas import get_schema_view
 
-from app.api import endpoints
-
+from app.api.doc import doc_endpoints
 
 urlpatterns = [
     # Auto Generated OpenAPI Schema
-    path('openapi-schema/', get_schema_view(
+    path('openapi/', get_schema_view(
         title="Red Mascotera API",
         description="API Endpoints for Red Mascotera",
         version="1.0.0"
     ), name='openapi-schema'),
     # Swagger View
-    path('swagger/', endpoints.SwaggerView.as_view(), name='swagger-ui'),
+    path('docs/', doc_endpoints.SwaggerView.as_view(), name='swagger-ui'),
 ]
