+++ red_mascotera/urls.py
@@ -19,5 +19,5 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('app.api.urls')),
+    path('api/', include('app.api.doc.doc_urls')),
 ]
